+++ test/profiling/integration.test.ts
@@ -8,8 +8,9 @@ import type { Envelope, Event, Profile, ThreadCpuProfile, Transaction, Transport
 
 import * as Sentry from '../../src/js';
 import { HermesProfiling } from '../../src/js/integrations';
+import type { NativeDeviceContextsResponse } from '../../src/js/NativeRNSentry';
 import { getDebugMetadata } from '../../src/js/profiling/debugid';
-import { isHermesEnabled } from '../../src/js/utils/environment';
+import { getDefaultEnvironment, isHermesEnabled } from '../../src/js/utils/environment';
 import { RN_GLOBAL_OBJ } from '../../src/js/utils/worldwide';
 import { MOCK_DSN } from '../mockDsn';
 import { envelopeItemPayload, envelopeItems } from '../testutils';
@@ -46,7 +47,7 @@ describe('profiling integration', () => {
   });
 
   test('should start profile if there is a transaction running when integration is created', () => {
-    mock = initTestClient(false);
+    mock = initTestClient({ withProfiling: false });
     jest.runAllTimers();
     jest.clearAllMocks();
 
@@ -61,6 +62,96 @@ describe('profiling integration', () => {
     jest.runAllTimers();
 
     expectEnveloperToContainProfile(mock.transportSendMock.mock.lastCall?.[0], 'test-name', transaction.traceId);
+  });
+
+  describe('environment', () => {
+    beforeEach(() => {
+      (getDefaultEnvironment as jest.Mock).mockReturnValue('mocked');
+      mockWrapper.NATIVE.fetchNativeDeviceContexts.mockResolvedValue(<NativeDeviceContextsResponse>{
+        environment: 'native',
+      });
+    });
+
+    const expectTransactionWithEnvironment = (envelope: Envelope | undefined, env: string | undefined) => {
+      const transactionEvent = envelope?.[envelopeItems][0][envelopeItemPayload] as Event;
+      expect(transactionEvent).toEqual(
+        expect.objectContaining<Partial<Event>>({
+          environment: env,
+        }),
+      );
+    };
+
+    const expectProfileWithEnvironment = (envelope: Envelope | undefined, env: string | undefined) => {
+      const profileEvent = (envelope?.[envelopeItems][1] as [{ type: 'profile' }, Profile])[1];
+      expect(profileEvent).toEqual(
+        expect.objectContaining<Partial<Profile>>({
+          environment: env,
+        }),
+      );
+    };
+
+    test('should use default environment for transaction and profile', () => {
+      mock = initTestClient();
+
+      const transaction: Transaction = Sentry.startTransaction({
+        name: 'test-name',
+      });
+      transaction.finish();
+
+      jest.runAllTimers();
+
+      const envelope: Envelope | undefined = mock.transportSendMock.mock.lastCall?.[0];
+      expectTransactionWithEnvironment(envelope, 'mocked');
+      expectProfileWithEnvironment(envelope, 'mocked');
+    });
+
+    test('should use native environment for transaction and profile if user value is nullish', () => {
+      mock = initTestClient({ withProfiling: true, environment: '' });
+
+      const transaction: Transaction = Sentry.startTransaction({
+        name: 'test-name',
+      });
+      transaction.finish();
+
+      jest.runAllTimers();
+
+      const envelope: Envelope | undefined = mock.transportSendMock.mock.lastCall?.[0];
+      expectTransactionWithEnvironment(envelope, 'native');
+      expectProfileWithEnvironment(envelope, 'native');
+    });
+
+    test('should keep nullish for transaction and profile uses default', () => {
+      mockWrapper.NATIVE.fetchNativeDeviceContexts.mockResolvedValue(<NativeDeviceContextsResponse>{
+        environment: undefined,
+      });
+      mock = initTestClient({ withProfiling: true, environment: undefined });
+
+      const transaction: Transaction = Sentry.startTransaction({
+        name: 'test-name',
+      });
+      transaction.finish();
+
+      jest.runAllTimers();
+
+      const envelope: Envelope | undefined = mock.transportSendMock.mock.lastCall?.[0];
+      expectTransactionWithEnvironment(envelope, undefined);
+      expectProfileWithEnvironment(envelope, 'mocked');
+    });
+
+    test('should keep custom environment for transaction and profile', () => {
+      mock = initTestClient({ withProfiling: true, environment: 'custom' });
+
+      const transaction: Transaction = Sentry.startTransaction({
+        name: 'test-name',
+      });
+      transaction.finish();
+
+      jest.runAllTimers();
+
+      const envelope: Envelope | undefined = mock.transportSendMock.mock.lastCall?.[0];
+      expectTransactionWithEnvironment(envelope, 'custom');
+      expectProfileWithEnvironment(envelope, 'custom');
+    });
   });
 
   describe('with profiling enabled', () => {
@@ -229,17 +320,24 @@ function getCurrentHermesProfilingIntegration(): TestHermesIntegration {
   return integration as unknown as TestHermesIntegration;
 }
 
-function initTestClient(withProfiling: boolean = true): {
+function initTestClient(
+  testOptions: {
+    withProfiling?: boolean;
+    environment?: string;
+  } = {
+    withProfiling: true,
+  },
+): {
   transportSendMock: jest.Mock<ReturnType<Transport['send']>, Parameters<Transport['send']>>;
 } {
   const transportSendMock = jest.fn<ReturnType<Transport['send']>, Parameters<Transport['send']>>();
-  Sentry.init({
+  const options: Sentry.ReactNativeOptions = {
     dsn: MOCK_DSN,
     _experiments: {
       profilesSampleRate: 1,
     },
     integrations: integrations => {
-      if (!withProfiling) {
+      if (!testOptions.withProfiling) {
         return integrations.filter(i => i.name !== 'HermesProfiling');
       }
       return integrations;
@@ -248,7 +346,11 @@ function initTestClient(withProfiling: boolean = true): {
       send: transportSendMock.mockResolvedValue(undefined),
       flush: jest.fn().mockResolvedValue(true),
     }),
-  });
+  };
+  if ('environment' in testOptions) {
+    options.environment = testOptions.environment;
+  }
+  Sentry.init(options);
 
   // In production integrations are setup only once, but in the tests we want them to setup on every init
   const integrations = Sentry.getCurrentHub().getClient()?.getOptions().integrations;